requirements.txt
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
streamlit
pyyaml

'''
#pip install -r requirements.txt

'''